Split review-voting components into separate files
